refactor: cleanup clipboard-history-view
